Allow passing a requests.Session to the Client
Using a requests.Session instead of the requests method directly gives
you a number of benefits such as connection pooling and allows you to
do more advanced connection configuration such as caching or playing
with adapters.

By allowing a session to be passed into the Client a user can configure
this behaviour without having to expose all this information via the
Client.

Signed-off-by: Jamie Lennox <jamielennox@gmail.com>